removed test components and udpated npm scripts
