+++ lib/client.js
@@ -17,7 +17,7 @@ var _spee = require("spee.ch-components");
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 // configure the reducers by passing initial state configs
-var siteConfig = require('./config/siteConfig.json'); // const customViews = require('./custom/views');
+var siteConfig = require('../config/siteConfig.json'); // const customViews = require('../custom/views');
 
 
 var MyReducers = (0, _spee.Reducers)(siteConfig);
