+++ lib/views/components/Logo/index.js
@@ -1,0 +1,18 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+
+var _react = _interopRequireDefault(require("react"));
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function Logo() {
+  return _react.default.createElement("p", null, "Test Logo");
+}
+
+;
+var _default = Logo;
+exports.default = _default;
