+++ lib/views/pages/AboutPage/index.js
@@ -7,7 +7,7 @@ exports.default = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
-var _spee = require("spee.ch-components");
+var _index = _interopRequireDefault(require("@containers/NavBar/index.js"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -29,7 +29,6 @@ function _assertThisInitialized(self) { if (self === void 0) { throw new Referen
 
 function _getPrototypeOf(o) { _getPrototypeOf = Object.getPrototypeOf || function _getPrototypeOf(o) { return o.__proto__; }; return _getPrototypeOf(o); }
 
-// import TestTest from 'test/TestTest';
 var AboutPage =
 /*#__PURE__*/
 function (_React$Component) {
@@ -42,7 +41,7 @@ function (_React$Component) {
   _createClass(AboutPage, [{
     key: "render",
     value: function render() {
-      return _react.default.createElement("div", null, _react.default.createElement(_spee.NavBar, null), _react.default.createElement("div", {
+      return _react.default.createElement("div", null, _react.default.createElement(_index.default, null), _react.default.createElement("div", {
         className: "row row--padded"
       }, _react.default.createElement("h3", null, "Welcome to your custom about page!"), _react.default.createElement("p", null, "This is an example of how you can create your own custom pages to replace the default spee.ch pages...")));
     }
