+++ package.json
@@ -4,14 +4,14 @@
   "description": "A web app that uses the spee.ch codebase",
   "main": "server.js",
   "scripts": {
-    "configure": "node cli/configure.js",
+    "babel": "babel src -d lib",
+    "babel:watch": "babel src -w -d lib",
     "build": "webpack --config webpack.config.js",
-    "build-dev": "webpack --config webpack.config.js --watch",
-    "build:lib": "babel src -d lib",
-    "build:lib:watch": "babel src -w -d lib",
+    "build:dev": "webpack --config webpack.config.js --watch",
+    "configure": "node cli/configure.js",
     "start": "node server.js",
-    "start-dev": "builder concurrent build-dev start-dev-server",
-    "start-dev-server": "nodemon server.js",
+    "start:dev": "builder concurrent babel:watch build:dev start:dev-server",
+    "start:dev-server": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
